# Fix a possible deadlock when Acquiring/Releasing different weights.