Hello,
so after further debugging for #3 
I found out that it actually has nothing to do with `GOMAXPROCS` as I assumed, it is just that different values`GOMAXPROCS` was affecting when releases happen in my example which made deadlocks less common when their values is greater.

The real issue happened when you don't use the same value for weights for acquiring and releasing **across the entire semaphore** .

The deadlock happened because the `Aquire` function didn't receive any signal on the broadcast channel.

Signal wasn't sent unless `count`(before release) >= `limit`, logically this assumes that a release of `n` will only notify waiters that are acquiring `n` too. the catch is if an Acquire of value (> `n`) is blocking it won't be notified by the releasing method because `count`(before release) < `limit` hence it assumes that no waiters are blocking. while in-fact an Acquire is blocking as ( `count`(before acquiring) + `weight` ) > `limit`.

The downside of this approach is that now a release is broadcast-ed on every release and another broadcast channel is allocated which will negatively affect performance.

### Benchmarks:
#### After:
```
goos: darwin
goarch: amd64
pkg: github.com/marusama/semaphore/bench
BenchmarkSemaphore_Acquire_Release_under_limit_simple-4   	20000000	       109 ns/op
BenchmarkSemaphore_Acquire_Release_under_limit-4          	  500000	      2589 ns/op
BenchmarkSemaphore_Acquire_Release_over_limit-4           	   50000	     27836 ns/op
PASS
```

#### Before:
```
goos: darwin
goarch: amd64
pkg: github.com/marusama/semaphore/bench
BenchmarkSemaphore_Acquire_Release_under_limit_simple-4   	50000000	        24.8 ns/op
BenchmarkSemaphore_Acquire_Release_under_limit-4          	 1000000	      1371 ns/op
BenchmarkSemaphore_Acquire_Release_over_limit-4           	  100000	     14299 ns/op
PASS
```

While this is unfortunate I think with the current implementation there is no way to avoid broadcasting releasing signal on every release unless weights is forced to be unified.

I've also included a test to that case when using different values for acquiring and releasing.