@Sherifabdlnaby 2019-01-09
Tests are passing but build is failing because of `The command "$HOME/gopath/bin/goveralls -coverprofile=coverage.out -service=travis-ci -repotoken $COVERALLS_TOKEN" exited with 2.`