@marusama 2019-01-10
Thanks for PR, @Sherifabdlnaby ! I also removed unnecessary comment as the condition has been removed. New semver release is 2.3.0, and also go.mod is added.